refactor: 홈페이지 구조 리펙토링
