+++ src/entities/notice/Notice.ts
@@ -1,0 +1,6 @@
+export interface Notice {
+  title: string;
+  date: string;
+  category: string;
+  important: boolean;
+}
