+++ src/features/notice/constants/RecentNotices.ts
@@ -1,0 +1,22 @@
+import type { Notice } from "@/entities/notice/Notice";
+
+export const recentNotices: Notice[] = [
+  {
+    title: "2025년 신년 감사예배 안내",
+    date: "2025.01.15",
+    category: "예배",
+    important: true,
+  },
+  {
+    title: "청년부 겨울 수련회 모집",
+    date: "2025.01.14",
+    category: "교육",
+    important: false,
+  },
+  {
+    title: "주일학교 교사 모집",
+    date: "2025.01.13",
+    category: "교육",
+    important: false,
+  },
+];
